Add more logging and remove asyncio
The asyncio.run call is not actually doing anything here. There is a call
to asyncio within the vespa sdk and the concern is that they could be
interfering with each other. I havent managed to prove this, so I'm also
adding a bunch of logging to get some extra information if the error comes
up again.