Ensure embeddings file is garbage collected
It turns out the embeddings file cache was not being cleared, presumably
because we where streaming the object before it was closed, resulting in
the file never getting closed, and therefore not being cleared of
cloudpathlibs cache. Testing on fixtures confirms this change results in
a clear cache.

Also removing a log line that effectively repeats another to make logs
easier to read through