Feature/pdct 692 speed up vespa indexing (#94)
* Finetune config for ~2x speed increase

Profiling different config on a subset of data led to these. Much higher
and vespa starts to creak at the seams and leads to errors. To offeset
this risk further I've also increased the reset countdown. The final
speed gain averaged 1.9125x faster with the new config

* Simplify task retrival function

This function had too many responsibilities. Also the multi line list
comprehensions added a lot of noise and made it harder to turn this into
a generator

* Load paths up front rather than whole documents

The task loader was parsing the json of every json file in the indexer
subdir. Instead this only loads the paths, opening each document when we
get to it. This should lower memory usage and will mean we start indexng
faster (and therefore surface issues faster)

* Simplify text block filter

Having moved this into the loop in the previous commit, this filter was
effectively running on one task at a time, this completes the refactor

* Move index path functions to utils

* Add test coverage for new functions

* Refactor existing tests following code refactor

* Clean away tqdm