Add logic to check for stray ids and clean them up
The problem with incremental runs is that occasionally a document will
get shorter. In other situations the old data gets overriden, but we need
to be able to handle the rare case that there will be less text blocks
in the new document as otherwise these will persist in vespa. Therefore,
this adds ligic to keep track of which ids remain and removes them once
the document has finished